# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Panier API Integration",
        "description": "API endpoints for communicating with the Panier API - ZIMRA Fiscalisation Support.\r\n\r\n## Rate Limit\r\n500 API calls every 5 minutes for each unique IP Address.\r\n\r\n## Authentication\r\nRequires APP-ID and API-KEY headers obtained from your Panier company account's Developer API section.\r\n\r\n## ZIMRA Fiscalisation\r\nThis API supports ZIMRA Fiscal Invoices, Credit Notes and Debit Notes based on ZIMRA Fiscal Device Gateway API Specs v7.2.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/quotation/create": {
            "post": {
                "tags": [
                    "Quotations"
                ],
                "summary": "Create quotations",
                "operationId": "9969394628462b05251d710b61f8f48e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/quotation/update": {
            "put": {
                "tags": [
                    "Quotations"
                ],
                "summary": "Update quotations",
                "operationId": "46dfb22495b1581a62b6099c93690077",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "id"
                                            ],
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/quotation/search": {
            "post": {
                "tags": [
                    "Quotations"
                ],
                "summary": "Search quotations",
                "operationId": "029f8766c3786fdbe5679f1e2cfd6c53",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/quotation/convert-to-invoice": {
            "post": {
                "tags": [
                    "Quotations"
                ],
                "summary": "Convert quotation to invoice",
                "operationId": "0bf976da97b9504970949784d2ee6c4f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "properties": {
                                            "quotation_id": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/quotation/convert-to-sale": {
            "post": {
                "tags": [
                    "Quotations"
                ],
                "summary": "Convert quotation to sale",
                "operationId": "797b4dc944f1ad22a7a216029f038e8e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "properties": {
                                            "quotation_id": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "zimra_fiscalize": {
                                        "type": "boolean",
                                        "default": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/quotation/download": {
            "get": {
                "tags": [
                    "Quotations"
                ],
                "summary": "Download quotation as PDF",
                "operationId": "2ab0c652cddab09b61ef8bdc81ea3506",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF file"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/quotation/delete": {
            "post": {
                "tags": [
                    "Quotations"
                ],
                "summary": "Delete quotations",
                "operationId": "a0f4cc93b3aa2fb049abbf47a6bbb9a1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "id"
                                            ],
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "AppId": {
                "type": "apiKey",
                "description": "Panier APP-ID from Developer API section",
                "name": "APP-ID",
                "in": "header"
            },
            "ApiKey": {
                "type": "apiKey",
                "description": "Panier API-KEY from Developer API section",
                "name": "API-KEY",
                "in": "header"
            }
        }
    },
...
}